Changed GraphSAGE and HinSAGE to take the base generator object.  * GraphSAGE/HinSAGE models require multiplicity argument if generator not passed  * HinSAGENodeGenerator requires a head_node_type argument * HinSAGELinkGenerator requires a head_node_types argument
