+++ stellargraph/data/explorer.py
@@ -20,6 +20,7 @@ __all__ = [
     "UniformRandomMetaPathWalk",
     "SampledBreadthFirstWalk",
     "SampledHeterogeneousBreadthFirstWalk",
+    "DirectedBreadthFirstNeighbours",
 ]
