+++ tests/layer/test_graph_attention.py
@@ -17,21 +17,16 @@
 """
 GAT tests
 """
-from stellargraph.core.graph import StellarGraph
-from stellargraph.mapper.node_mappers import (
-    FullBatchNodeGenerator,
-    GraphSAGENodeGenerator,
-)
-from stellargraph.layer import *
-
-from tensorflow import keras
-from tensorflow.keras import backend as K
-from tensorflow.keras.layers import Input
-
 import numpy as np
 import networkx as nx
 import pytest
 import scipy.sparse as sps
+from tensorflow import keras
+from tensorflow.keras import backend as K
+from tensorflow.keras.layers import Input
+from stellargraph.core.graph import StellarGraph
+from stellargraph.mapper import FullBatchNodeGenerator, GraphSAGENodeGenerator
+from stellargraph.layer import *
 
 
 def example_graph_1(feature_size=None):
